Data: Accession numbers and metadata of 97 E. coli genomes from Malawi from Musciha et al

Metadata of a collection of 97 E. coli genomes selected for temporal diversity and diversity in antibiogram Taken from Musicha et al 2017 For full details see original Musicha publication https://doi.org/10.1093/jac/dkx058

btESBL_ecoli_musicha_metadata

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 97 rows and 10 columns.

Details

The phylogroup and ST here were determined for this study and were not provided in the original metadata; the popPUNK cluster assignemt uses the Horesh database

#'

lane

Unique sample-sequencing run ID

ID

Unmique sample ID

Year

Year of sample collection

Sample_type

Type of clincial sample from which isolate originates

Age_group

Age of patient from which sample orignated

Organism

Organism

ST

Multlocus sequence type as determined by ARIBA

phylogroup

Phylogroup following Clermon scheme with in silico PCR

Cluster

PopPUNK cluster assignment using the Horesh E. coli database

Accession

NCBI Accession number