Data: Exposures of included participants

These data are recorded in long format. Covariate values are included in columns, and there is one row for each day where a covariate value changes or when the participant has their final observation (i.e date of censoring). So, if a participant has a covariate value that changes from 1 to 0 on subsequent rows, the covariate value should be interpreted as being 1 up until the day of change

btESBL_exposures

Format

A data frame with 1838 rows and 26 variables:

pid

Unique patient identifier

assess_type

Day of observation (day 0=enrolment day)

died

1=died on day of current observation

hosp

1=hospitalised on day of current observation, 0 =not

amoxy

Exposure to amoxicillin on day of observation

genta

Exposure to gentamicin on day of observation

azithro

Exposure to azithromycin on day of observation

tb

Exposure to first-line TB treatment on day of observation

benzy

Exposure to penicillin on day of observation

fluco

Exposure to fluconazole on day of observation

ampho

Exposure to amphotericin on day of observation

quin

Exposure to quinine on day of observation

chlora

Exposure to chloraquine on day of observation

coart

Exposure to artemether/lumefantrine on day of observation

cipro

Exposure to ciprofloxacin on day of observation

coamo

Exposure to co-amoxiclav on day of observation

arte

Exposure to artesunate on day of observation

cotri

Exposure to co-trimoxazole on day of observation

clinda

Exposure to clindamicin on day of observation

doxy

Exposure to doxycycline on day of observation

erythro

Exposure to erythomycin on day of observation

acicl

Exposure to aciclovir on day of observation

fluclox

Exposure to flucloxacillin on day of observation

metro

Exposure to metronidazole on day of observation

strepto

Exposure to streptomicin on day of observation