Data: Results of simulations from the model posterior

These simulations set the antimicrobial and hospital exposure to arbitrary values, set starting probability of colonisation to 0.5 then draw parameter values from the posterior of the fitted model (btESBL_model2posterior) to generate probability of being ESBL colonised a time t later

btESBL_model2simulations

Format

A data frame with 600,000 rows and 18 variables:

time

Number of days post enrollment

1

Probability of being ESBL non-colonised at current time

0

Probability of being ESBL colonised at current time

draw

Unique ID of draw from posterior (1,2...1000)

hosp_days

Number of days of hospitalisation

abx_days

Number of days of antimicrobial exposure

p0

Probability at time 0 of being uncolonised

p1

Probability at time 0 of being colonised

abx_start

Time of antimicrobial exposure start

abx_stop

Time of antimicrobial exposure stop

prev_abx

Time of cessation of prev antimicrobial exposure (999=none)

hosp_start

Time of hospitalisation start

hosp_stop

Time of hospitalisation stop

prev_hosp

Time of cessation of prev hospitalisation (999=none)